requirements.txt
asgiref==3.2.7
dj-config-url==0.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-money==1.1
psycopg2-binary==2.8.4
py-moneyed==0.8.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
static3==0.7.0
gunicorn==20.0.4
python-dateutil==2.8.1
pynubank==1.2.0
wait-for==1.1.1
django-colorfield==0.3.0
